# speedup add_simplices_from