`add_simplices_from` is slow, see #221. Profiling showed about 90% of the time is spent in the line checking existence of the simplex: 
```python
if not members or self.has_simplex(members):
```

`has_simplex()` now access members through `S._edge` rather than `S.edges.members(s)` which is significantly faster (see #221 for timings).

Here, adding 100 triangles seems about 10x faster:

```python
%timeit S.add_simplices_from(list(triangles)[:100]) 
# old -> 9.13 ms ± 28.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# new -> 860 µs ± 5.18 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

The `add_simplices_from` can probably be improved even more, by rewriting it it without recursion for example. I have written a version but not pushed it because the seped improvement seems negligible in my case (it might not be with larger simplices). Also, it modifies the method much more, so has a larger chance of new bugs. And it changes the order in which simplices are added, so tests would need to be udpated. I would start with this minimal update at first, and maybe use the other version for a second step.