@maximelucas 2022-11-11
Removed unnecessary generator expression in `has_simplex`, yielding further speed up (thanks profiling).

Timings, wrt to previous speedup. For
```python
S = xgi.SimplicialComplex()
simplices = [{0, 1, 2, 3}, {4}, {5, 6}, {6, 7, 8}]
```

```python
%timeit S.add_simplices_from(simplices, max_order=2)
# old -> 7.45 µs ± 10.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
# new -> 4.72 µs ± 16.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

```python 
%timeit xgi.random_simplicial_complex(250, [0.1, 0.001])
# old -> 2.92 s ± 65.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# new -> 1.72 s ± 42.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

```python
%timeit S.add_simplices_from(triangles) # 61712 triangles
# old -> 2min 20s ± 331 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# new -> 1min 17s ± 801 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

So we get an additional 2x speedup (especially when adding more simplices).